pyproject.toml
[project]
name = "fpg-api"
version = "0.1.0"
description = "This Repo is the link between the FPG Database and the FPG front end"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "flasgger>=0.9.7.1",
    "flask>=3.1.0",
    "flask-httpauth>=4.8.0",
    "mysql-connector-python>=9.3.0",
    "pandas>=2.2.3",
    "pytest>=8.3.5",
    "sqlalchemy>=2.0.40",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["test_api"]

